uncrustify source code and make C++11 compatible

+++ geometry/s2cellid.cc
@@ -216,7 +216,7 @@ S2CellId S2CellId::FromFaceIJ(int face, int i, int j) {
   // rather than local variables helps the compiler to do a better job
   // of register allocation as well.  Note that the two 32-bits halves
   // get shifted one bit to the left when they are combined.
-  uint32 n[2] = { 0, face << (kPosBits - 33) };
+  uint32 n[2] = { 0, static_cast<uint32>(face << (kPosBits - 33)) };
 
   // Alternating faces have opposite Hilbert curve orientations; this
   // is necessary in order for all faces to have a right-handed
